Refactor test models into internal/testing/testdata
