+++ internal/testing/testdata/models.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package expected
+package testdata
 
 import (
 	"github.com/guacsec/guac/internal/testing/ptrfrom"
